+++ kplayer/online_example/flutter_service_worker.js
@@ -3,23 +3,22 @@ const MANIFEST = 'flutter-app-manifest';
 const TEMP = 'flutter-temp-cache';
 const CACHE_NAME = 'flutter-app-cache';
 const RESOURCES = {
-  "assets/AssetManifest.json": "e760b19954155f009ca0c475a33586d2",
+  "assets/AssetManifest.json": "e06719576acdfb9af3b5f332e875486e",
 "assets/assets/Introducing_flutter.mp3": "32e35c93843649dc0a526c41083ffed6",
-"assets/FontManifest.json": "dc3d03800ccca4601324923c0b1d6d57",
+"assets/FontManifest.json": "7b2a36307916a9721811788013e65289",
 "assets/fonts/MaterialIcons-Regular.otf": "95db9098c58fd6db106f1116bae85a0b",
-"assets/NOTICES": "24df87c499f7d2d46676fa76c783061a",
-"assets/packages/cupertino_icons/assets/CupertinoIcons.ttf": "6d342eb68f170c97609e9da345464e5e",
+"assets/NOTICES": "eac82c9c47c75f86afdf01dfb6f11caa",
 "canvaskit/canvaskit.js": "c2b4e5f3d7a3d82aed024e7249a78487",
 "canvaskit/canvaskit.wasm": "4b83d89d9fecbea8ca46f2f760c5a9ba",
 "canvaskit/profiling/canvaskit.js": "ae2949af4efc61d28a4a80fffa1db900",
 "canvaskit/profiling/canvaskit.wasm": "95e736ab31147d1b2c7b25f11d4c32cd",
 "favicon.png": "5dcef449791fa27946b3d35ad8803796",
-"flutter.js": "3688efe0a39e59781b4f95efbd6b5b62",
+"flutter.js": "0816e65a103ba8ba51b174eeeeb2cb67",
 "icons/Icon-192.png": "ac9a721a12bbc803b44f645561ecb1e1",
 "icons/Icon-512.png": "96e752610906ba2a93c65f8abe1645f1",
-"index.html": "97cf9ef50c48f241fb43e453a40c0f72",
-"/": "97cf9ef50c48f241fb43e453a40c0f72",
-"main.dart.js": "ecc8b7b3ba770dfcc8f1feb5dd4d5688",
+"index.html": "025a0184f952b2a1316fc61c919c1557",
+"/": "025a0184f952b2a1316fc61c919c1557",
+"main.dart.js": "0883bc8204aee1ab59ac7eeb5a7b166c",
 "manifest.json": "5c4fffb56495d51b28801e7b41ef6a67",
 "version.json": "a13b211dbe24d6219935f7b6b20f12e3"
 };