+++ kplayer/kplayer_with_audioplayers/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.2.0
+
+check kplayer package for more information
 ## 0.0.1
 
 * TODO: Describe initial release.